ml_fastapi/requirements.txt
tensorflow-cpu==2.17.0
scikit-learn
fastapi
uvicorn[standard]
python-dotenv
requests
pandas
numpy
python-multipart